+++ ceammc/gui/plugins/ceammc_popup_menu_tree.tcl
@@ -200,8 +200,8 @@
         synth.risset_arp~ synth.risset_tone~ synth.shakers~ synth.wurley~
     }}
     {system {
-        system.cursor system.getenv system.hostname system.memsize
-        system.memused system.screen_size system.shell
+        system.colorpanel system.cursor system.getenv system.hostname
+        system.memsize system.memused system.screen_size system.shell
     }}
     {tl {
         tl.bang tl.cue tl.timeline tl.toggle tl.transport